+++ mitmproxy/addons/savehar.py
@@ -2,30 +2,51 @@
 import base64
 import json
 import logging
+import zlib
 from collections.abc import Sequence
 from datetime import datetime
 from datetime import timezone
 from typing import Any
 
 from mitmproxy import command
+from mitmproxy import ctx
+from mitmproxy import exceptions
 from mitmproxy import flow
+from mitmproxy import flowfilter
 from mitmproxy import http
 from mitmproxy import types
 from mitmproxy import version
 from mitmproxy.connection import Server
 from mitmproxy.coretypes.multidict import _MultiDict
-from mitmproxy.exceptions import CommandError
+from mitmproxy.log import ALERT
+from mitmproxy.utils import human
 from mitmproxy.utils import strutils
 
 logger = logging.getLogger(__name__)
 
 
-class ExportHar:
-    @command.command("exporthar")
+class SaveHar:
+    def __init__(self) -> None:
+        self.flows: list[flow.Flow] = []
+        self.filt: flowfilter.TFilter | None = None
+
+    @command.command("save.har")
     def export_har(self, flows: Sequence[flow.Flow], path: types.Path) -> None:
         """Export flows to an HAR (HTTP Archive) file."""
-        entries = []
 
+        har = json.dumps(self.make_har(flows), indent=4).encode()
+
+        if path.endswith(".zhar"):
+            har = zlib.compress(har, 9)
+
+        with open(path, "wb") as f:
+            f.write(har)
+
+        logging.log(ALERT, f"HAR file saved ({human.pretty_size(len(har))} bytes).")
+
+    def make_har(self, flows: Sequence[flow.Flow]) -> dict:
+        entries = []
+        skipped = 0
         # A list of server seen till now is maintained so we can avoid
         # using 'connect' time for entries that use an existing connection.
         servers_seen: set[Server] = set()
@@ -34,22 +55,73 @@ class ExportHar:
             if isinstance(f, http.HTTPFlow):
                 entries.append(self.flow_entry(f, servers_seen))
             else:
-                raise CommandError(f"Cannot export {type(f).__name__} flows to HAR.")
+                skipped += 1
 
-        har = {
+        if skipped > 0:
+            logger.info(f"Skipped {skipped} flows that weren't HTTP flows.")
+
+        return {
             "log": {
                 "version": "1.2",
                 "creator": {
-                    "name": "mitmproxy exporthar",
-                    "version": "0.1",
-                    "comment": f"mitmproxy version {version.VERSION}",
+                    "name": "mitmproxy",
+                    "version": version.VERSION,
+                    "comment": "",
                 },
                 "pages": [],
                 "entries": entries,
             }
         }
-        with open(path, "w") as fp:
-            json.dump(har, fp, indent=4)
+
+    def load(self, l):
+        l.add_option(
+            "hardump",
+            str,
+            "",
+            "Save a HAR file with all flows on exit. "
+            "You may select particular flows by setting save_stream_filter.",
+            "For mitmdump, enabling this option will mean that flows are kept in memory. ",
+        )
+
+    def configure(self, updated):
+        if "save_stream_filter" in updated:
+            if ctx.options.save_stream_filter:
+                try:
+                    self.filt = flowfilter.parse(ctx.options.save_stream_filter)
+                except ValueError as e:
+                    raise exceptions.OptionsError(str(e)) from e
+            else:
+                self.filt = None
+
+        if "hardump" in updated:
+            if not ctx.options.hardump:
+                self.flows = []
+
+    def response(self, flow: http.HTTPFlow) -> None:
+        # websocket flows will receive a websocket_end,
+        # we don't want to persist them here already
+        if flow.websocket is None:
+            self._save_flow(flow)
+
+    def error(self, flow: http.HTTPFlow) -> None:
+        self.response(flow)
+
+    def websocket_end(self, flow: http.HTTPFlow) -> None:
+        self._save_flow(flow)
+
+    def _save_flow(self, flow: http.HTTPFlow) -> None:
+        if ctx.options.hardump:
+            flow_matches = self.filt is None or self.filt(flow)
+            if flow_matches:
+                self.flows.append(flow)
+
+    def done(self):
+        if ctx.options.hardump:
+            if ctx.options.hardump == "-":
+                har = self.make_har(self.flows)
+                print(json.dumps(har, indent=4))
+            else:
+                self.export_har(self.flows, ctx.options.hardump)
 
     def flow_entry(self, flow: http.HTTPFlow, servers_seen: set[Server]) -> dict:
         """Creates HAR entry from flow"""
@@ -89,6 +161,7 @@ class ExportHar:
             receive = 1000 * (
                 flow.response.timestamp_end - flow.response.timestamp_start
             )
+
         else:
             receive = 0
 
@@ -178,6 +251,23 @@ class ExportHar:
         if flow.server_conn.peername:
             entry["serverIPAddress"] = str(flow.server_conn.peername[0])
 
+        websocket_messages = []
+        if flow.websocket:
+            for message in flow.websocket.messages:
+                if message.is_text:
+                    data = message.text
+                else:
+                    data = base64.b64encode(message.content).decode()
+                websocket_message = {
+                    "type": "send" if message.from_client else "receive",
+                    "time": message.timestamp,
+                    "opcode": message.type.value,
+                    "data": data,
+                }
+                websocket_messages.append(websocket_message)
+
+            entry["_resourceType"] = "websocket"
+            entry["_webSocketMessages"] = websocket_messages
         return entry
 
     def format_response_cookies(self, response: http.Response) -> list[dict]:
@@ -189,7 +279,7 @@ class ExportHar:
                 "name": name,
                 "value": value,
                 "path": attrs["path"],
-                "domain": attrs["domain"],
+                "domain": attrs.get("domain", ""),
                 "httpOnly": "httpOnly" in attrs,
                 "secure": "secure" in attrs,
             }
@@ -205,6 +295,3 @@ class ExportHar:
 
     def format_multidict(self, obj: _MultiDict[str, str]) -> list[dict]:
         return [{"name": k, "value": v} for k, v in obj.items(multi=True)]
-
-
-addons = [ExportHar()]